requirements.txt
Django
djangorestframework
psycopg2-binary
celery
redis
gunicorn
python-dotenv
pytest
pytest-django
